add trees

+++ SimWorld/OpenGL/LTree/TreeBranch.h
@@ -33,7 +33,7 @@
 /// <summary>
 /// Index of the parent branch, or -1 if this is the root branch.
 /// </summary>
-@property (atomic) int parentIndex;
+@property (atomic) NSUInteger parentIndex;
 
 /// <summary>
 /// Where on the parent the branch is located. 0.0f is at the start, 1.0f is at the end.
@@ -43,7 +43,7 @@
 /// <summary>
 /// Index of the bone controlling this branch.
 /// </summary>
-@property (atomic) int boneIndex;
+@property (atomic) NSUInteger boneIndex;
 
 - (id)initWithQuaternion:(CC3GLMatrix *)rotation
                andLength:(float)length
